Keyboard speed control in playback direction, FX chain active in pause mode, fixed playback of audio, fixed audio up-and downsampling (like speeding up analogue tape), fixed recording of audio (if video sample has audio), fixed slot motion playback, increased AVI file limit to 32 gb, fixed preview rendering, cleanup of audio playback and recording, gveejayreloaded: sample bank pictures dont mix up anymore, button 'slow motion multiplier' added to sample panel, disable 'veejay framerate' when playing audio
git-svn-id: svn://code.dyne.org/veejay/trunk@766 eb8d1916-c9e9-0310-b8de-cf0c9472ead5

+++ veejay-current/libel/lav_io.h
@@ -102,7 +102,7 @@ int  lav_seek_start(lav_file_t *lav_file);
 int  lav_set_video_position(lav_file_t *lav_file, long frame);
 int  lav_read_frame(lav_file_t *lav_file, uint8_t *vidbuf);
 int  lav_set_audio_position(lav_file_t *lav_file, long clip);
-long lav_read_audio(lav_file_t *lav_file, uint8_t *audbuf, long samps);
+int  lav_read_audio(lav_file_t *lav_file, uint8_t *audbuf, long samps);
 int  lav_filetype(lav_file_t *lav_file);
 lav_file_t *lav_open_input_file(char *filename, int mmap_size);
 int  lav_get_field_size(uint8_t * jpegdata, long jpeglen);
